requirements.txt
django
python-decouple
graphene-django
pyjwt
django-cors-headers
gunicorn
psycopg2-binary
dj-database-url
whitenoise
bcrypt
stripe